JPA Annotations + model cleanup

+++ src/main/java/com/pkaushik/safeHome/repository/RequestRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RequestRepository extends JpaRepository<SpecificRequest, Integer> {
+public interface RequestRepository extends JpaRepository<SpecificRequest, Long> {
 
 }
